backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.9
pydantic==2.7.0
tiktoken==0.7.0
sentence-transformers==3.0.1
scikit-learn==1.5.0
numpy==1.26.4
groq==0.9.0
google-generativeai==0.7.2
pymupdf==1.24.5
python-docx==1.1.2
python-pptx==0.6.23
openpyxl==3.1.5
Pillow==10.4.0
python-dotenv==1.0.1
slowapi==0.1.9